backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0

# Anthropic SDK
anthropic>=0.52.0

# File handling
aiofiles>=24.1.0
python-multipart>=0.0.19

# HTTP client (CLI, batch orchestrator, testing)
httpx>=0.27.0

# Trusted monitor (transcript evaluation)
# pip install git+https://github.com/safety-research/trusted-monitor.git
trusted-monitor>=0.1.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24
